Problem: pipeline files cannot be loaded via glob
Earlier configurations expected all machines
to reside in the same parent directory and so
pipeline files could be found by filename and
loaded. This is no longer the case and pipeling
now is part of the lings package. Use an rpc-like
approach.

Solution: remove pipline file search and import code