Problem: lings-pipe-* cli's do not work as expected
If the command is run with a simple positional, the
expectation is that it should work. Flags can be added
for specificity, but the simplest case should work such
as lings-pipe-get some_name

Solution: add handling for simple use case/expectation.
If a positional is supplied as first arg it is assumed to
be the name/uuid to be used by cli.