Problem: no way to pub from arbitrary functions
Since messages published on channels serve
as starting points for routes, it would be
useful to easily specify places and messages
for publishing without inserting code inside
of every function.

Solution: use a python decorator that also does
some formatting so that argument names can be
passed in as curly braced strings and correct value
will be inserted.